refactor: split image canvas logic
